% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake-driver.R
\name{dbDataType,SnowflakeConnection-method}
\alias{dbDataType,SnowflakeConnection-method}
\title{Find the database data type associated with an R object}
\usage{
\S4method{dbDataType}{SnowflakeConnection}(dbObj, obj, ...)
}
\description{
Find the database data type associated with an R object
}
