% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake-driver.R
\name{dbFetch,SnowflakeResult-method}
\alias{dbFetch,SnowflakeResult-method}
\title{Retrieve records from Kazam query}
\usage{
\S4method{dbFetch}{SnowflakeResult}(res, n = -1, ...)
}
\description{
Retrieve records from Kazam query
}
