% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake-driver.R
\name{dbSendQuery,SnowflakeConnection-method}
\alias{dbSendQuery,SnowflakeConnection-method}
\title{Send a query to Snowflake}
\usage{
\S4method{dbSendQuery}{SnowflakeConnection}(conn, statement, head, ...)
}
\description{
Send a query to Snowflake
}
