% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowcaseSeedling.R
\name{Display}
\alias{Display}
\title{Display}
\usage{
Display(tree, reflections = 5, linecolor = "#6BAED6")
}
\arguments{
\item{tree}{slope of the line}

\item{reflections}{the number of times the tree is to be reflected around the origin}

\item{linecolor}{the desired color of the tree}
}
\description{
Display a snowflake from the fully grown plant produced by `Grow`.
}
\author{
Daniel R. Kick (\email{drk8b9@mail.missouri.edu})
}
