% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowcaseSeedling.R
\name{Rotate}
\alias{Rotate}
\title{Rotate}
\usage{
Rotate(SegmentDF, d = 180)
}
\arguments{
\item{SegmentDF}{A daframe with columns x, y, xend, and yend}

\item{d}{The angle to rotate points in degrees}
}
\description{
Given a dataframe with the minimum mapping for `geom_segment`
(x, y, xend, yend) and a degree of rotation, rotate all points
by the given degree around the origin.
}
\author{
Daniel R. Kick (\email{drk8b9@mail.missouri.edu})
}
