% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowSeedling.R
\name{Step}
\alias{Step}
\title{Step}
\usage{
Step(x, y, d, m)
}
\arguments{
\item{x}{starting x}

\item{y}{starting y}

\item{d}{direction in degrees}

\item{m}{magnitude of step}
}
\description{
Given starting coordinates, a direction in degrees,
and a magnitude, return the coordinates for the vector's end.
}
\author{
Daniel R. Kick (\email{drk8b9@mail.missouri.edu})
}
