% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowSeedling.R
\name{Water}
\alias{Water}
\title{Water}
\usage{
Water(Stem = Stem, d = temp[i, "d"], m = temp[i, "m"], r = temp[i, "r"])
}
\arguments{
\item{Stem}{list object produced within `Grow` from Guidelines. Contains the previous segment which is being branched.}

\item{d}{direction in degrees}

\item{m}{magnitude of step}

\item{r}{regression from end}
}
\description{
This function opperates withing `Grow`. It takes an active stem, finds
where the new branches will split off (with `r`), then returns two new
stems `d` degrees to either side of the parent stem  with length `m`.
}
\author{
Daniel R. Kick (\email{drk8b9@mail.missouri.edu})
}
